refactor: reuse block that adds scripts
